+++ pit/internal/version/version_test.go
@@ -3,7 +3,7 @@ package version
 import "testing"
 
 func TestString(t *testing.T) {
-	if String() != "PIT 0.6.0" {
+	if String() != "PIT 0.6.1" {
 		t.Fatal(String())
 	}
 }